Added Comments, Updated MLFQ for LONG jobs.
MLFQ, while it worked for the inputs given, would have run into a problem with very long jobs because it would have tried to move down to a priority of 6 (index 5), which would have likely segfaulted the program. The queue should round-robin better when it gets down to 4, as well.